% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcFC.R
\name{calcFC}
\alias{calcFC}
\title{Calculate a forecast from a single statistical model}
\usage{
calcFC(
  fit.obj = NULL,
  data = NULL,
  data.sibreg = NULL,
  fc.yr = NULL,
  settings = NULL,
  tracing = FALSE,
  predictors = NULL,
  covariates = NULL
)
}
\arguments{
\item{fit.obj}{A list, created by \code{\link{fitModel}}, with model fitting
details for each age class.}

\item{data}{A list. Equal to the element named \code{data} from output of
\code{\link{prepData}}.}

\item{data.sibreg}{A data frame. Equal to the element named \code{sibreg.in} from output of
\code{\link{prepData}}. NOTE: all the functions and subroutines should be rewritten to just use the full \code{\link{prepData}} output, but priority now is to get the new sibreg variations working}

\item{fc.yr}{An integer of length one. The forecast year. If NULL (the default), then calculate as the year after the last run year in the data set.}

\item{settings}{A list. The model-specific list of settings used in \code{\link{fitModel}}. NOTE: This is now part of the list object created by \code{\link{fitModel}}, so should streamline the input.}

\item{tracing}{A Boolean. Default is FALSE}

\item{predictors}{A list. Optional input needed for return rate model. Equal to the element named \code{predictors} from output of
\code{\link{prepData}}. Default is NULL.}

\item{covariates}{A list. Optional input needed for covariate model. Equal to the element named \code{covariates} from output of
\code{\link{prepData}}. Default is NULL.}
}
\value{

}
\description{
Calculate a forecast from a single statistical model
}
