% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Sub_preSibRegData.R
\name{getSets}
\alias{getSets}
\title{get all possible sets of pooled sibreg predictor variables}
\usage{
getSets(x)
}
\arguments{
\item{x}{a continuous vector of age classes in the data set}
}
\description{
get all possible sets of pooled sibreg predictor variables
}
