% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Sub_preSibRegData.R
\name{prepSibRegData}
\alias{prepSibRegData}
\title{generate single df with all individual age data and pooled sibreg inputs lined up by brood year}
\usage{
prepSibRegData(data.by.age)
}
\arguments{
\item{data.by.age}{this is the $data element of the prepData() output,called tmpsub internally}
}
\description{
generate single df with all individual age data and pooled sibreg inputs lined up by brood year
}
