% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_transformDF.R
\name{transformDF}
\alias{transformDF}
\title{apply the transformVar() function to all numeric columns in a data frame}
\usage{
transformDF(X, type, logfirst = FALSE)
}
\arguments{
\item{X}{is a data frame to transform using \code{\link{transformVar}}.}

\item{type}{"none", "standardize" (rescale to 0-1), or "normalize" (rescale to mean=0, sd =1)}

\item{logfirst}{if TRUE, do a log-transform BEFORE applying either standardize or normalize}
}
\value{
A data frame with numeric columns transformed
}
\description{
apply the transformVar() function to all numeric columns in a data frame
}
\details{
If any zeroes in the data, then the log-transform causes problems. 
Using the strategy from Perry et al 2021 
\link{(https://journals.plos.org/plosone/article/comments?id=10.1371/journal.pone.0245941)}
which replaces 0 with a random value greater than 0 and less than one-half of the
lowest non-zero value in the data for that variable. If the data has negative values, a 
constant equal to min(x)* 1.01 is added to all values (i.e. smallest value is a 
bit larger than 0)
}
\examples{
transformDF(X = mtcars, type = "standardize", logfirst = FALSE)
transformDF(X = mtcars, type = "normalize", logfirst = TRUE)
}
