% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_transformDF.R
\name{transformVar}
\alias{transformVar}
\title{transform a numeric vector}
\usage{
transformVar(x, type, logfirst = FALSE, warning.print = FALSE)
}
\arguments{
\item{x}{is a numeric vector}

\item{type}{"none", "standardize" (rescale to 0-1), or "normalize" (rescale to mean=0, sd =1)}

\item{logfirst}{if TRUE, do a log-transform BEFORE applying either standardize or normalize}

\item{warning.print}{if TRUE print a warning if Perry et al 2021 0 fix is triggered}
}
\value{
A transformed vector
}
\description{
transform a numeric vector
}
\details{
Typically called from within  \code{\link{transformDF}}. If any zeroes in the data,
then the log-transform causes problems. 
Using the strategy from Perry et al 2021 
\link{(https://journals.plos.org/plosone/article/comments?id=10.1371/journal.pone.0245941)}
which replaces 0 with a random value greater than 0 and less than one-half of the
lowest non-zero value in the data for that variable. If the data has negative values, a 
constant equal to min(x)* 1.01 is added to all values (i.e. smallest value is a 
bit larger than 0).
}
\examples{
transformVar(X = mtcars$mpg, type = "standardize", logfirst = FALSE)
transformVarX = mtcars$mpg, type = "normalize", logfirst = TRUE)
}
